Force process list refresh every 20 ticks.
Because of the following check we actually had to do this every tick.

  if (procs_count != NUM_INTERESTING_PROCS) {
      procs_stale = 1;
  }

Initially I expected that NUM_INTERESTING_PROCS would always be equal
to the number of child processes. So the check would force the refresh
only when we knew that were some processes missing that potentially
could start later. But then because of some differences on Windows we
just increased NUM_INTERESTING_PROCS. And hence we actually had to
reread the list of processes on every tick. AFAIR, we've seen signs of
that in some of the customer diags.

To avoid this behavior from now we just force periodic update.

Change-Id: I90a49f805230424f6533043537e3f3a9f5fb42f5
Reviewed-on: http://review.couchbase.org/46908
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>