MB-32205: Fix heap-buffer-overflow in find_interesting_procs
When searching for 'interesting' processes, sigar looks for child
processes matching a given parent process, and iterates across all
child processes. However there is a potential heap overflow during the
iteration - it can dereference past the end of the array (procs_end)
via child->ppid, as it checks for reaching the end of the array
*after* dereferencing the pointer.

Fix by flipping the order of the check.

Change-Id: I0e146407a06c0d8df06403a26a0d875b9468e8df
Reviewed-on: http://review.couchbase.org/102328
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Build Bot <build@couchbase.com>