Take babysitter pid as a command line parameter.
Erlang 20 starts all processes via an intermediate process called
erl_child_setup. This breaks the old logic of finding out the
babysitter pid by going two levels up in the process hierarchy. And as
a consequence of this, we don't collect information about processes
like memcached.

This change addresses this by accepting the babysitter pid on the
command line instead of trying to deduce it.

Change-Id: Ia1e103f5a40ff6205a8feaf9ea85fcefe7689300
Reviewed-on: http://review.couchbase.org/112378
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>