MB-34865: let sigar_os_proc_list_get only return my processes
sigar_port wants information about the child processes of the
babysitter, but the API to sigar won't let it just query just
that. Instead it needs to walk over all processes and look at
the parent processes (but in order to fetch the ppid it collects
a lot of other stuff).

In sigar_os_proc_list_get we've got both the ppid and the uid
available, but to avoid changing too much in the code I figured
we could skip all process _not_ belonging to ourselves and then
hopefully get rid of any access restrictions later on.

Change-Id: I0d38ce5b91c471cac51e951a50e4062870c2ae30
Reviewed-on: http://review.couchbase.org/113948
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>