Don't leave gaps in replies when skipping processes.
The old code would always write the process information at the same
index as the process in the list of interesting processes. But the
loop could skip some of the processes (for instance, when a process
has died). That would leave gaps in the reply. So a separate
index for the replies needs to be maintained.

Change-Id: I565627f2adad8bc5266c48b1be1ba06d88a9a26d
Reviewed-on: http://review.couchbase.org/122821
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>