Fix process cpu info staleness threshold.
If the comment around this code to be trusted, it was meant to be 1
second. In reality, it was 1 millisecond.

We do occasionally call sigar_proc_cpu_get on the same process in
quick succession. With the bad threshold, if less than 1 millisecond
passed between the calls, we'd get a cached garbage value for cpu
utilization (fixed by the preceding patch). Otherwise, we'd likely get
a zero (I can see that execution time stays the same in these
situations, but I'm not sure if it's the kernel that doesn't update
the information frequently enough, or if it's sigar caching the
information somewhere). This patch bumps the threshold to the intended
value of 1 second. Even with the fix the behavior is not perfect, but
it more or less works with how we use sigar.

Change-Id: Iaf782789f0e0a92a2a8d8ca74c239bfb58abb916
Reviewed-on: http://review.couchbase.org/122823
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>