MB-38296 Report /proc/vmstat's allocstall stat
Sigar will return the following information from /proc/vmstat on linux
systems.

up until 4.10
        allocstall,             /* up until 4.10 */

4.10 onwards the sum of the following
        allocstall_dma,         /* 4.10 onwards */
        allocstall_dma32,       /* 4.10 onwards */
        allocstall_normal,      /* 4.10 onwards */
        allocstall_movable;     /* 4.10 onwards */

If the stat isn't available and on non-linux systems  it will return -1.

allocstall is the number of times direct reclaim occurs.

From https://lwn.net/Articles/396561/

Normally, memory pages are reclaimed (made available for new uses, with
data written back, if necessary) in the background; when all goes well,
there will always be a list of free pages available when memory is needed If,
however, a memory allocation request cannot be satisfied from the free list,
the kernel may try to reclaim pages directly in the context of the process
performing the allocation. Diverting an allocation request into this kind of
cleanup activity is called "direct reclaim."

This change requires the following ns_server change:
http://review.couchbase.org/#/c/124033/

Change-Id: I8a5be6d5a372e64f457b9cff69ac47a64cf2619e
Reviewed-on: http://review.couchbase.org/123991
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>