Remove caching of pid
It isn't much point of caching the process pid as all it does
is to shave off a system call, and introduce the problem with
fork() not resetting it correctly. Linux handled the case, but
macos did not.

THe library don't call the method itself, so the user could
always cache it themselves if they don't want the overhead
of the system call

Change-Id: I4d044c06782a027b4f706cb347a5b8541500cfde
Reviewed-on: https://review.couchbase.org/c/sigar/+/166394
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>