MB-49911: Make sure we don't overflow buffers
sigar_file2str tries to fill a buffer with data and finally
add a termination to the buffer.

The problem arise when we fill the entire buffer we will add
the termination outside the buffer.

Ex:
   char buffer[8192]
   int len = read(fd, buffer, sizeof(buffer));
   buffer[len] = '\0';

If read returns 8192 we will try to set the byte outside
the allocated array.

Change-Id: Ie0466c309845cb8dcc4a83d8aedb9805b2b82b00
Reviewed-on: https://review.couchbase.org/c/sigar/+/167034
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>