MB-52653: Report allocstall even if one entry is missing
On Amazon Linux 2 - Graviton instance /proc/vmstat don't contain
the same entries as on lets say Ubuntu on x86. sigar_port
would only report allocstall if it:

a) had an entry named allocstall
b) had all of the entries found on x86.

This patch sums up the entries we know about if some of
them was found. If none of them was found we return
not implemented (-1)

Graviton instance:
    grep allocstall /proc/vmstat
    allocstall_dma 0
    allocstall_normal 0
    allocstall_movable 0

Ubuntu 22.04LTS on x86:
    grep allocstall /proc/vmstat
    allocstall_dma 0
    allocstall_dma32 0
    allocstall_normal 81
    allocstall_movable 103

Change-Id: Ia6281a798f7eb0f0b113c12ad54c111e2d760c6e
Reviewed-on: https://review.couchbase.org/c/sigar/+/176579
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>