MB-55513: allow sending JSON output
Enable by passing --json on the command line.

New command line options are:

   --json[=pretty]
   --babysitter_pid=<pid>
   --help

When running in JSON mode the "trigger" to get the next sample
is the newline character. Closing the input stream (or sending
'quit\n') will cause sigar_port to terminate. The motivation
behind this change is to allow easy testing on the command
line (without having to write 4 bytes with the integer value
0)

Change-Id: Ie95bd15980a805e2167c31d580fa05661bdac85d
Reviewed-on: https://review.couchbase.org/c/sigar/+/186495
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>