[MacOS] Remove unused branch from get_proc_memory
The implementation of get_proc_memory first tried to use
proc_pidinfo to read out the process information, and if that
failed it would fall back to use task_info to pick the
information. The first method should not fail if we have
access to the process, and if that fails the second
method will _also_ fail.

Change-Id: I9861c6baed0966fa14a7790f41adf0bc183bd26a
Reviewed-on: https://review.couchbase.org/c/sigar/+/187422
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>