Allow for requesting a snapshot
When invoked from the command line you typically just
want to see the result. It is a bit annoying to have
to write:

    echo "next\nquit\n" | ./sigar_port --json

Add a new option --snapshot which will just print out
a single sample and then terminate.

When running on a tty we don't need to write out the
size of the JSON payload as the user just want to
see the output.

Change-Id: If2933e0b1c5225267b545c92eaf6d2a647029975
Reviewed-on: https://review.couchbase.org/c/sigar/+/187486
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>