MB-55680: Fix read of vm.pagesize
We passed in 8 bytes, but the value was reported as a 4 byte
value. Fixed this by first requesting the size of the variable
and then read it with the correct size.

In addition to that we can cache the total memory and page
size as it won't change for the lifetime of the process.

Change-Id: I1e7c156bd0cb476d35d927f76cc2e99778e4ea7b
Reviewed-on: https://review.couchbase.org/c/sigar/+/187493
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>