MB-56411: Add support for log file for sigar_port
To ease troubleshooting on customers environments sigar_port
should log errors to its own file and use unique exit codes.

sigar_port should now be started as:

    sigar_port --babysitter_pid <pid> \
               --logfile <filename> \
               --config <filename>

This is a reimplementation of the patch from 7.2.1 as the
code have changed significantly between the two versions.
This version utilize spdlog to write the log.

Change-Id: I9b23f7e247e5939cb26da6a2759ac44b84c98e0b
Reviewed-on: https://review.couchbase.org/c/sigar/+/191282
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>