MB-59236: Check perf configuration on startup
During initialization on windows sigar will try to look up
the various perf providers to check if they're disabled, and
if so print a log message. Given that the system refuse to
start when there is a problem, it's easier to have customers
run:

   sigar_port --human-readable --snapshot --babysitter_pid=self

And see the error information written to the terminal (and possibly
fix the problems and rerun the command line).

Change-Id: I09ef1a7cdc19274dd5c24371557252046ed82dd6
Reviewed-on: https://review.couchbase.org/c/sigar/+/199624
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>