MB-59554: [win32] Don't use perf for process info [1/4]
All the information related to process info may be fetched
without using perf, and instead of falling back to using that
API when we're encountering an error with PerfProc this patch
adds implementation to look up the process memory with
GetProcessMemoryInfo.

The patch returns "no process" if it fails to open the process
for some reason (The process don't exists, or if there is a
problem with the access rights for instance). The actual error
message will be logged (except if the return value indicates
that the process don't exists). The motivation behind returning
no process for other errors is that if this was indeed an access
permission problem the caller we don't want the information
anyway (our current usage only wants to inspect processes running
by the same user anyway).

If the system calls for fetching memory usage, parent pid,
thread count or executable name fails, their values would
be set to "not implemented" (-1) (the executable would be named
"unknown").

Change-Id: I65febb78fbb27250ec5f41487ebee5326f655388
Reviewed-on: https://review.couchbase.org/c/sigar/+/200235
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>